Move {address,subnet}_index into separate files
