+++ libvast/src/value_index.cpp
@@ -148,195 +148,6 @@ caf::error inspect(caf::deserializer& source, value_index_ptr& x) {
   return x->deserialize(source);
 }
 
-// -- address_index ------------------------------------------------------------
-
-address_index::address_index(vast::type t, caf::settings opts)
-  : value_index{std::move(t), std::move(opts)} {
-  bytes_.fill(byte_index{8});
-}
-
-caf::error address_index::serialize(caf::serializer& sink) const {
-  return caf::error::eval([&] { return value_index::serialize(sink); },
-                          [&] { return sink(bytes_, v4_); });
-}
-
-caf::error address_index::deserialize(caf::deserializer& source) {
-  return caf::error::eval([&] { return value_index::deserialize(source); },
-                          [&] { return source(bytes_, v4_); });
-}
-
-bool address_index::append_impl(data_view x, id pos) {
-  auto addr = caf::get_if<view<address>>(&x);
-  if (!addr)
-    return false;
-  auto& bytes = addr->data();
-  for (auto i = 0u; i < 16; ++i) {
-    bytes_[i].skip(pos - bytes_[i].size());
-    bytes_[i].append(bytes[i]);
-  }
-  v4_.skip(pos - v4_.size());
-  v4_.append(addr->is_v4());
-  return true;
-}
-
-caf::expected<ids>
-address_index::lookup_impl(relational_operator op, data_view d) const {
-  return caf::visit(
-    detail::overload{
-      [&](auto x) -> caf::expected<ids> {
-        return make_error(ec::type_clash, materialize(x));
-      },
-      [&](view<address> x) -> caf::expected<ids> {
-        if (!(op == equal || op == not_equal))
-          return make_error(ec::unsupported_operator, op);
-        auto result = x.is_v4() ? v4_.coder().storage() : ids{offset(), true};
-        for (auto i = x.is_v4() ? 12u : 0u; i < 16; ++i) {
-          auto bm = bytes_[i].lookup(equal, x.data()[i]);
-          result &= bm;
-          if (all<0>(result))
-            return ids{offset(), op == not_equal};
-        }
-        if (op == not_equal)
-          result.flip();
-        return result;
-      },
-      [&](view<subnet> x) -> caf::expected<ids> {
-        if (!(op == in || op == not_in))
-          return make_error(ec::unsupported_operator, op);
-        auto topk = x.length();
-        if (topk == 0)
-          return make_error(ec::unspecified,
-                            "invalid IP subnet length: ", topk);
-        auto is_v4 = x.network().is_v4();
-        if ((is_v4 ? topk + 96 : topk) == 128)
-          // Asking for /32 or /128 membership is equivalent to an equality
-          // lookup.
-          return lookup_impl(op == in ? equal : not_equal, x.network());
-        auto result = is_v4 ? v4_.coder().storage() : ids{offset(), true};
-        auto& bytes = x.network().data();
-        size_t i = is_v4 ? 12 : 0;
-        for (; i < 16 && topk >= 8; ++i, topk -= 8)
-          result &= bytes_[i].lookup(equal, bytes[i]);
-        for (auto j = 0u; j < topk; ++j) {
-          auto bit = 7 - j;
-          auto& bm = bytes_[i].coder().storage()[bit];
-          result &= (bytes[i] >> bit) & 1 ? ~bm : bm;
-        }
-        if (op == not_in)
-          result.flip();
-        return result;
-      },
-      [&](view<list> xs) { return detail::container_lookup(*this, op, xs); },
-    },
-    d);
-}
-
-// -- subnet_index -------------------------------------------------------------
-
-subnet_index::subnet_index(vast::type x, caf::settings opts)
-  : value_index{std::move(x), std::move(opts)},
-    network_{address_type{}},
-    length_{128 + 1} {
-  // nop
-}
-
-caf::error subnet_index::serialize(caf::serializer& sink) const {
-  return caf::error::eval([&] { return value_index::serialize(sink); },
-                          [&] { return sink(network_, length_); });
-}
-
-caf::error subnet_index::deserialize(caf::deserializer& source) {
-  return caf::error::eval([&] { return value_index::deserialize(source); },
-                          [&] { return source(network_, length_); });
-}
-
-bool subnet_index::append_impl(data_view x, id pos) {
-  if (auto sn = caf::get_if<view<subnet>>(&x)) {
-    length_.skip(pos - length_.size());
-    length_.append(sn->length());
-    return static_cast<bool>(network_.append(sn->network(), pos));
-  }
-  return false;
-}
-
-caf::expected<ids>
-subnet_index::lookup_impl(relational_operator op, data_view d) const {
-  return caf::visit(
-    detail::overload{
-      [&](auto x) -> caf::expected<ids> {
-        return make_error(ec::type_clash, materialize(x));
-      },
-      [&](view<address> x) -> caf::expected<ids> {
-        if (!(op == ni || op == not_ni))
-          return make_error(ec::unsupported_operator, op);
-        auto result = ids{offset(), false};
-        uint8_t bits = x.is_v4() ? 32 : 128;
-        for (uint8_t i = 0; i <= bits; ++i) { // not an off-by-one
-          auto masked = x;
-          masked.mask(128 - bits + i);
-          ids len = length_.lookup(equal, i);
-          auto net = network_.lookup(equal, masked);
-          if (!net)
-            return net;
-          len &= *net;
-          result |= len;
-        }
-        if (op == not_ni)
-          result.flip();
-        return result;
-      },
-      [&](view<subnet> x) -> caf::expected<ids> {
-        switch (op) {
-          default:
-            return make_error(ec::unsupported_operator, op);
-          case equal:
-          case not_equal: {
-            auto result = network_.lookup(equal, x.network());
-            if (!result)
-              return result;
-            auto n = length_.lookup(equal, x.length());
-            *result &= n;
-            if (op == not_equal)
-              result->flip();
-            return result;
-          }
-          case in:
-          case not_in: {
-            // For a subnet index U and subnet x, the in operator signifies a
-            // subset relationship such that `U in x` translates to U ⊆ x, i.e.,
-            // the lookup returns all subnets in U that are a subset of x.
-            auto result = network_.lookup(in, x);
-            if (!result)
-              return result;
-            *result &= length_.lookup(greater_equal, x.length());
-            if (op == not_in)
-              result->flip();
-            return result;
-          }
-          case ni:
-          case not_ni: {
-            // For a subnet index U and subnet x, the ni operator signifies a
-            // subset relationship such that `U ni x` translates to U ⊇ x, i.e.,
-            // the lookup returns all subnets in U that include x.
-            ids result;
-            for (auto i = uint8_t{1}; i <= x.length(); ++i) {
-              auto xs = network_.lookup(in, subnet{x.network(), i});
-              if (!xs)
-                return xs;
-              *xs &= length_.lookup(equal, i);
-              result |= *xs;
-            }
-            if (op == not_ni)
-              result.flip();
-            return result;
-          }
-        }
-      },
-      [&](view<list> xs) { return detail::container_lookup(*this, op, xs); },
-    },
-    d);
-}
-
 // -- port_index ---------------------------------------------------------------
 
 port_index::port_index(vast::type t, caf::settings opts)
