+++ libvast/src/value_index_factory.cpp
@@ -19,9 +19,11 @@
 #include "vast/detail/bit.hpp"
 #include "vast/detail/type_traits.hpp"
 #include "vast/hash_index.hpp"
+#include "vast/index/address_index.hpp"
 #include "vast/index/arithmetic_index.hpp"
 #include "vast/index/enumeration_index.hpp"
 #include "vast/index/string_index.hpp"
+#include "vast/index/subnet_index.hpp"
 #include "vast/logger.hpp"
 #include "vast/type.hpp"
 #include "vast/value_index.hpp"
