+++ libvast/vast/index/subnet_index.hpp
@@ -1,0 +1,55 @@
+/******************************************************************************
+ *                    _   _____   __________                                  *
+ *                   | | / / _ | / __/_  __/     Visibility                   *
+ *                   | |/ / __ |_\ \  / /          Across                     *
+ *                   |___/_/ |_/___/ /_/       Space and Time                 *
+ *                                                                            *
+ * This file is part of VAST. It is subject to the license terms in the       *
+ * LICENSE file found in the top-level directory of this distribution and at  *
+ * http://vast.io/license. No part of VAST, including this file, may be       *
+ * copied, modified, propagated, or distributed except according to the terms *
+ * contained in the LICENSE file.                                             *
+ ******************************************************************************/
+
+#pragma once
+
+#include "vast/bitmap_index.hpp"
+#include "vast/coder.hpp"
+#include "vast/error.hpp"
+#include "vast/ewah_bitmap.hpp"
+#include "vast/ids.hpp"
+#include "vast/index/address_index.hpp"
+#include "vast/value_index.hpp"
+#include "vast/view.hpp"
+
+#include <caf/error.hpp>
+#include <caf/expected.hpp>
+#include <caf/fwd.hpp>
+
+#include <array>
+#include <cstdint>
+
+namespace vast {
+
+/// An index for subnets.
+class subnet_index : public value_index {
+public:
+  using prefix_index = bitmap_index<uint8_t, equality_coder<ewah_bitmap>>;
+
+  explicit subnet_index(vast::type t, caf::settings opts = {});
+
+  caf::error serialize(caf::serializer& sink) const override;
+
+  caf::error deserialize(caf::deserializer& source) override;
+
+private:
+  bool append_impl(data_view x, id pos) override;
+
+  caf::expected<ids>
+  lookup_impl(relational_operator op, data_view x) const override;
+
+  address_index network_;
+  prefix_index length_;
+};
+
+} // namespace vast
